Tortue implantée + changement en considération des classes alphabets et seDessiner

+++ src/alphabet/Mouvement.java
@@ -3,9 +3,9 @@ import java.util.LinkedList;
 
 
 public abstract class Mouvement extends MembreAlpha {
-	protected int valeur;
+	protected double valeur;
 	protected boolean tourne;
-	public Mouvement(Character representation, LinkedList<MembreAlpha> evolution ,int valeur) {
+	public Mouvement(Character representation, LinkedList<MembreAlpha> evolution ,double valeur) {
 		super(representation,evolution);
 		this.valeur=valeur;
 		this.dessine=false;
@@ -15,14 +15,14 @@ public abstract class Mouvement extends MembreAlpha {
 		return tourne;
 	}
 
-	public Mouvement(Character representation, int valeur) {
+	public Mouvement(Character representation, double valeur) {
 		super(representation);
 		this.valeur=valeur;
 	}
-	public int getValeur() {
+	public double getValeur() {
 		return valeur;
 	}
-	public void setValeur( int valeur) {
+	public void setValeur( double valeur) {
 		this.valeur = valeur;
 	}
